Add return link to demo homepage
