+++ my_website/my_website/views/_static/home_controller-7005a669dca6d9a486b169db2b9b0005.js
@@ -509,7 +509,7 @@ var require_react_development = __commonJS({
           }
           return element;
         };
-        function createElement3(type, config, children) {
+        function createElement2(type, config, children) {
           var propName;
           var props = {};
           var key = null;
@@ -1608,7 +1608,7 @@ var require_react_development = __commonJS({
               error("React.createElement: type is invalid -- expected a string (for built-in components) or a class/function (for composite components) but got: %s.%s", typeString, info);
             }
           }
-          var element = createElement3.apply(this, arguments);
+          var element = createElement2.apply(this, arguments);
           if (element == null) {
             return element;
           }
@@ -9387,7 +9387,7 @@ var require_react_dom_development = __commonJS({
             }
           }
         }
-        function createElement3(type, props, rootContainerElement, parentNamespace) {
+        function createElement2(type, props, rootContainerElement, parentNamespace) {
           var isCustomComponentTag;
           var ownerDocument = getOwnerDocumentFromRootContainer(rootContainerElement);
           var domElement;
@@ -10248,7 +10248,7 @@ var require_react_dom_development = __commonJS({
             }
             parentNamespace = hostContextDev.namespace;
           }
-          var domElement = createElement3(type, props, rootContainerInstance, parentNamespace);
+          var domElement = createElement2(type, props, rootContainerInstance, parentNamespace);
           precacheFiberNode(internalInstanceHandle, domElement);
           updateFiberProps(domElement, props);
           return domElement;
@@ -23506,25 +23506,87 @@ var require_client = __commonJS({
   }
 });
 
-// ../../../../../../private/var/folders/0z/txmshp9s1679jxprrlw8f8_h0000gn/T/filzl-buildtAHrBV/synthetic_client.tsx
+// ../../../../../../private/var/folders/0z/txmshp9s1679jxprrlw8f8_h0000gn/T/filzl-buildgOZh2A/synthetic_client.tsx
 var React3 = __toESM(require_react());
 var import_client = __toESM(require_client());
 
-// ../../my_website/my_website/views/app/detail/page.tsx
-var React2 = __toESM(require_react());
+// ../../my_website/my_website/views/app/home/page.tsx
+var import_react2 = __toESM(require_react());
 
-// ../../my_website/my_website/views/app/detail/_server/useServer.ts
+// ../../my_website/my_website/views/app/home/_server/useServer.ts
 var import_react = __toESM(require_react());
 
-// ../../my_website/my_website/views/app/home/_server/links.ts
-var getLink = ({
-  home_id
-}) => {
-  let url = "/home/{home_id}/";
-  const queryParameters = {};
-  const pathParameters = {
-    home_id
+// ../../my_website/my_website/views/_server/api.ts
+var FetchErrorBase = class extends Error {
+  statusCode;
+  body;
+  constructor(statusCode, body) {
+    super(`Error ${statusCode}: ${body}`);
+    this.statusCode = statusCode;
+    this.body = body;
+  }
+};
+var handleOutputFormat = async (response, format) => {
+  if (format === "text") {
+    return await response.text();
+  } else {
+    return await response.json();
+  }
+};
+var __request = async (params) => {
+  const payloadBody = params.body ? JSON.stringify(params.body) : void 0;
+  let filledUrl = params.url;
+  for (const [key, value] of Object.entries(params.path || {})) {
+    filledUrl = filledUrl.replace(`{${key}}`, value.toString());
+  }
+  Object.entries(params.query || {}).forEach(([key, value], i) => {
+    filledUrl = `${filledUrl}${i === 0 ? "?" : "&"}${key}=${value}`;
+  });
+  try {
+    const response = await fetch(filledUrl, {
+      method: params.method,
+      headers: {
+        "Content-Type": params.mediaType || "application/json"
+      },
+      body: payloadBody
+    });
+    if (response.status >= 200 && response.status < 300) {
+      return await handleOutputFormat(response, params.outputFormat);
+    } else {
+      if (params.errors && params.errors[response.status]) {
+        const errorClass = params.errors[response.status];
+        throw new errorClass(
+          response.status,
+          await handleOutputFormat(response, params.outputFormat)
+        );
+      }
+      throw new FetchErrorBase(
+        response.status,
+        await handleOutputFormat(response, params.outputFormat)
+      );
+    }
+  } catch (e) {
+    if (e instanceof FetchErrorBase) {
+      throw e;
+    }
+    const error = new FetchErrorBase(-1, e.toString());
+    error.stack = e.stack;
+    throw error;
+  }
+};
+function applySideEffect(apiFunction, setControllerState) {
+  return async (...args) => {
+    const result = await apiFunction(...args);
+    setControllerState(result.sideeffect);
+    return result;
   };
+}
+
+// ../../my_website/my_website/views/app/home/_server/links.ts
+var getLink = ({}) => {
+  let url = "/";
+  const queryParameters = {};
+  const pathParameters = {};
   const parsedParams = Object.entries(queryParameters).reduce((acc, [key, value]) => {
     if (value !== void 0) {
       acc.push(`${key}=${value}`);
@@ -23579,7 +23641,55 @@ var linkGenerator = {
 };
 var links_default = linkGenerator;
 
-// ../../my_website/my_website/views/app/detail/_server/useServer.ts
+// ../../my_website/my_website/views/app/home/_server/actions.ts
+var get_external_data = () => {
+  return __request(
+    {
+      "method": "POST",
+      "url": "/internal/api/home_controller/get_external_data",
+      "query": {}
+    }
+  );
+};
+var increment_count = ({
+  requestBody
+}) => {
+  return __request(
+    {
+      "method": "POST",
+      "url": "/internal/api/home_controller/increment_count",
+      "query": {},
+      "errors": {
+        422: HTTPValidationErrorException
+      },
+      "body": requestBody,
+      "mediaType": "application/json"
+    }
+  );
+};
+var increment_count_only = ({
+  url_param,
+  requestBody
+}) => {
+  return __request(
+    {
+      "method": "POST",
+      "url": "/internal/api/home_controller/increment_count_only",
+      "query": {
+        url_param
+      },
+      "errors": {
+        422: HTTPValidationErrorException
+      },
+      "body": requestBody,
+      "mediaType": "application/json"
+    }
+  );
+};
+var HTTPValidationErrorException = class extends FetchErrorBase {
+};
+
+// ../../my_website/my_website/views/app/home/_server/useServer.ts
 var useServer = () => {
   const [serverState, setServerState] = (0, import_react.useState)(SERVER_DATA);
   const setControllerState = (payload) => {
@@ -23590,18 +23700,56 @@ var useServer = () => {
   };
   return {
     ...serverState,
-    linkGenerator: links_default
+    linkGenerator: links_default,
+    get_external_data,
+    increment_count: applySideEffect(increment_count, setControllerState),
+    increment_count_only: applySideEffect(increment_count_only, setControllerState)
   };
 };
 
-// ../../my_website/my_website/views/app/detail/page.tsx
-var Page = () => {
+// ../../my_website/my_website/views/app/home/page.tsx
+var Home = () => {
   const serverState = useServer();
-  return /* @__PURE__ */ React2.createElement("div", null, /* @__PURE__ */ React2.createElement("h1", null, "Server"), /* @__PURE__ */ React2.createElement("pre", null, JSON.stringify(serverState, null, 2)));
+  console.log("SERVER PAYLOAD", serverState);
+  return /* @__PURE__ */ import_react2.default.createElement("div", null, /* @__PURE__ */ import_react2.default.createElement("h1", null, "Home"), /* @__PURE__ */ import_react2.default.createElement("p", null, "Home page"), /* @__PURE__ */ import_react2.default.createElement("p", null, "Hello ", serverState.client_ip, ", current count is", " ", serverState.current_count), /* @__PURE__ */ import_react2.default.createElement(
+    "a",
+    {
+      href: serverState.linkGenerator.detailController({
+        detail_id: "9280ca4d-e607-486a-a343-7bbaae6b5a86"
+      })
+    },
+    "Detail Link"
+  ), /* @__PURE__ */ import_react2.default.createElement(
+    "button",
+    {
+      onClick: async () => {
+        await serverState.increment_count({
+          requestBody: {
+            count: 1
+          }
+        });
+      }
+    },
+    "Increment"
+  ), /* @__PURE__ */ import_react2.default.createElement(
+    "button",
+    {
+      onClick: async () => {
+        await serverState.increment_count_only({
+          // Not used, but demonstrates that it's possible to pass a url param
+          url_param: 5,
+          requestBody: {
+            count: -1
+          }
+        });
+      }
+    },
+    "Decrement with sideeffect masking"
+  ));
 };
-var page_default = Page;
+var page_default = Home;
 
-// ../../../../../../private/var/folders/0z/txmshp9s1679jxprrlw8f8_h0000gn/T/filzl-buildtAHrBV/synthetic_client.tsx
+// ../../../../../../private/var/folders/0z/txmshp9s1679jxprrlw8f8_h0000gn/T/filzl-buildgOZh2A/synthetic_client.tsx
 var Entrypoint = () => {
   return /* @__PURE__ */ React3.createElement(page_default, null);
 };
@@ -23655,4 +23803,4 @@ react-dom/cjs/react-dom.development.js:
    * @license Modernizr 3.0.0pre (Custom Build) | MIT
    *)
 */
-//# sourceMappingURL=detail_controller-45a138deccf7f468a89007fb0725a784.js.map
+//# sourceMappingURL=home_controller-7005a669dca6d9a486b169db2b9b0005.js.map
