+++ my_website/my_website/views/app/detail/_server/links.ts
@@ -1,33 +1,36 @@
-export const getLink = ({ detail_id }: { detail_id: string }) => {
-  let url = "/detail/{detail_id}/";
+export const getLink = ({
+detail_id
+} : {
+detail_id: string
+}) => {
+let url = '/detail/{detail_id}/';
 
-  const queryParameters: Record<string, string> = {};
-  const pathParameters: Record<string, string> = {
-    detail_id,
-  };
+const queryParameters : Record<string, string> = {
 
-  const parsedParams = Object.entries(queryParameters).reduce(
-    (acc, [key, value]) => {
-      if (value !== undefined) {
-        acc.push(`${key}=${value}`);
-      }
-      return acc;
-    },
-    [] as string[],
-  );
+};
+const pathParameters : Record<string, string> = {
+detail_id
+};
 
-  const paramString = parsedParams.join("&");
+const parsedParams = Object.entries(queryParameters).reduce((acc, [key, value]) => {
+if (value !== undefined) {
+acc.push(`${key}=${value}`);
+}
+return acc;
+}, [] as string[]);
 
-  for (const [key, value] of Object.entries(pathParameters)) {
-    if (value === undefined) {
-      throw new Error(`Missing required path parameter ${key}`);
-    }
-    url = url.replace(`{${key}}`, value);
-  }
+const paramString = parsedParams.join('&');
 
-  if (paramString) {
-    url = `${url}?${paramString}`;
-  }
+for (const [key, value] of Object.entries(pathParameters)) {
+if (value === undefined) {
+throw new Error(`Missing required path parameter ${key}`);
+}
+url = url.replace(`{${key}}`, value);
+}
 
-  return url;
+if (paramString) {
+url = `${url}?${paramString}`;
+}
+
+return url;
 };
