+++ my_website/my_website/views/app/home/_server/actions.ts
@@ -1,5 +1,5 @@
 import { __request, FetchErrorBase } from '../../../_server/api';
-import type { IncrementCountOnlyResponse, HTTPValidationError, IncrementCountRequest, IncrementCountResponse, GetExternalDataResponse } from './models';
+import type { HTTPValidationError, IncrementCountResponse, IncrementCountOnlyResponse, IncrementCountRequest, GetExternalDataResponse } from './models';
 
 export const get_external_data = (): Promise<GetExternalDataResponse> => {
 return __request(
